\name{drugData}
\alias{drugData}
\docType{data}
\title{Drug Data Title}
\description{Drug data description}
\usage{data(drugData)}
\details{
  The activity data.
}
\source{Website link}
\value{
  the drugData object that contains drug information as a DrugData object
}
\references{PMID: 0000}
\concept{rcellminerDemoData}
\keyword{datasets}
