\name{molData}
\alias{molData}
\docType{data}
\title{Molecular Data Title}
\description{Molecular data description}
\usage{data(molData)}
\details{
Data matrices for the following types of data are provided. NOTE: Missing values are represented as NA. 

  \itemize{
    \item cop: DESCRIPTION
    \item mut: DESCRIPTION
    \item exp: DESCRIPTION
  }
}
\source{Website link}
\value{
  the molData object that contains molecular profiling information as a MolData object
}
\references{PMID: 0000}
\concept{rcellminerDemoData}
\keyword{datasets}
